% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpipe.R
\name{backpipe}
\alias{backpipe}
\title{backpipe}
\usage{
backpipe(pipe, backpipe = gsub(">", "<", pipe))
}
\arguments{
\item{pipe}{character; string representing the existing pipe operator}

\item{backpipe}{character; string representing the desired backpipe operator}
}
\description{
Creates backpiping operators
}
\details{
Only \code{pipe} is necessary. Arbitrary mixing of forward and reverse are
not allowed.
}
\examples{
  backpipe('\%>\%')
  backpipe('\%>>\%', '\%<<\%')
                    
}
\references{
\url{https://github.com/smbache/magrittr/issues/26} \cr
\url{http://stackoverflow.com/questions/31305342/is-right-to-left-operator-associativity-in-r-possible} \cr
}
\seealso{
\code{\link[magrittr]{\%>\%}} \cr
\code{\link[pipeR]{\%>>\%}} \cr
}
