% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_backpipe.R
\name{insert_backpipe}
\alias{insert_backpipe}
\title{Insert backpipe}
\usage{
insert_backpipe()
}
\description{
Call this function to insert \code{\%&lt;-\%} at the cursor position.
}
\details{
Use \strong{Tools > Modify Keyboard Shortcuts} to bind a  ->.;  ->.;
}
