% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_backpipe_addin.R
\name{insert_backpipe_addin}
\alias{insert_backpipe_addin}
\title{Backpipe Shortcut}
\usage{
insert_backpipe_addin()
}
\description{
Call this function as an addin to insert \code{\%&lt;-\%} at the cursor position.
}
\details{
Use \strong{Tools > Modify Keyboard Shortcuts} to bind a  ->.;  ->.;
}
